+++ src/main/java/domain/game/TakeMoreCardSelector.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.game;
 
+import domain.card.Card;
 import java.util.List;
 
 public interface TakeMoreCardSelector {
