+++ src/main/java/dto/FinalResultDTO.java
@@ -1,5 +1,6 @@
-package domain;
+package dto;
 
+import domain.card.Card;
 import java.util.List;
 
 public record FinalResultDTO(String name, List<Card> cards, int score) {
