+++ src/main/java/dto/SetUpCardsDTO.java
@@ -1,5 +1,6 @@
-package domain;
+package dto;
 
+import domain.card.Card;
 import java.util.List;
 import java.util.Map;
