+++ src/test/java/domain/game/GameManagerTest.java
@@ -1,9 +1,14 @@
-package domain;
+package domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Shape;
 import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
