+++ src/test/java/domain/game/WinningTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
